+++ core/lib/Cardano/Codec/Bech32/Prefixes.hs
@@ -70,19 +70,25 @@ module Cardano.Codec.Bech32.Prefixes
     , drep_sk
     , drep_xvk
     , drep_xsk
+    , drep_vkh
     , drep
+    , drep_script
 
     , cc_cold_vk
     , cc_cold_sk
     , cc_cold_xvk
     , cc_cold_xsk
+    , cc_cold_vkh
     , cc_cold
+    , cc_cold_script
 
     , cc_hot_vk
     , cc_hot_sk
     , cc_hot_xvk
     , cc_hot_xsk
+    , cc_hot_vkh
     , cc_hot
+    , cc_hot_script
     ) where
 
 import Codec.Binary.Bech32
@@ -250,8 +256,14 @@ drep_xvk = [humanReadablePart|drep_xvk|]
 drep_xsk :: HumanReadablePart
 drep_xsk = [humanReadablePart|drep_xsk|]
 
+drep_vkh :: HumanReadablePart
+drep_vkh = [humanReadablePart|drep_vkh|]
+
 drep :: HumanReadablePart
 drep = [humanReadablePart|drep|]
+
+drep_script :: HumanReadablePart
+drep_script = [humanReadablePart|drep_script|]
 
 cc_cold_vk :: HumanReadablePart
 cc_cold_vk = [humanReadablePart|cc_cold_vk|]
@@ -262,11 +274,17 @@ cc_cold_sk = [humanReadablePart|cc_cold_sk|]
 cc_cold_xvk :: HumanReadablePart
 cc_cold_xvk = [humanReadablePart|cc_cold_xvk|]
 
+cc_cold_vkh :: HumanReadablePart
+cc_cold_vkh = [humanReadablePart|cc_cold_vkh|]
+
 cc_cold_xsk :: HumanReadablePart
 cc_cold_xsk = [humanReadablePart|cc_cold_xsk|]
 
 cc_cold :: HumanReadablePart
 cc_cold = [humanReadablePart|cc_cold|]
+
+cc_cold_script :: HumanReadablePart
+cc_cold_script = [humanReadablePart|cc_cold_script|]
 
 cc_hot_vk :: HumanReadablePart
 cc_hot_vk = [humanReadablePart|cc_hot_vk|]
@@ -280,5 +298,11 @@ cc_hot_xvk = [humanReadablePart|cc_hot_xvk|]
 cc_hot_xsk :: HumanReadablePart
 cc_hot_xsk = [humanReadablePart|cc_hot_xsk|]
 
+cc_hot_vkh :: HumanReadablePart
+cc_hot_vkh = [humanReadablePart|cc_hot_vkh|]
+
 cc_hot :: HumanReadablePart
 cc_hot = [humanReadablePart|cc_hot|]
+
+cc_hot_script :: HumanReadablePart
+cc_hot_script = [humanReadablePart|cc_hot_script|]